Merged InvoiceListener and InvoiceBindingListener
They now are one unified InvoiceListener, with events fired
by the binding. This alleviates callers from the burden of
registering both an invoice listener and an invoice binding
listener in order to know when the open invoice changed so
as to unregister and re-register the invoice listener. With
this unified interface, callers can just register one listener
and asve a bunch of boilerplate.